requirements.txt
streamlit
pandas
streamlit-autorefresh
gspread
google-auth